Refactor app into modules and fix missing imports
